refactor: add model computer add repository for computer
